Control menu principal de galaxias funcionanado (Menu superior)
