+++ app/src/main/AndroidManifest.xml
@@ -41,6 +41,10 @@
             android:name=".levels.ExerciseVersionThree"
             android:exported="false">
         </activity>
+        <activity
+            android:name=".ProfileActivity"
+            android:exported="false">
+        </activity>
     </application>
 
 </manifest>
