+++ app/src/main/java/com/rcs_rocket_english/ProfileActivity.java
@@ -1,0 +1,13 @@
+package com.rcs_rocket_english;
+
+import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class ProfileActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.profile);
+    }
+}
